% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_funcs.R
\name{add_slash_if_not}
\alias{add_slash_if_not}
\title{Add slash to directory name if not already there}
\usage{
add_slash_if_not(d)
}
\arguments{
\item{d}{A path to a directory}
}
\value{
The directory location string, finished by a \code{"/"}.
}
\description{
Adds a slash to a directory location if it doesn't already finish by a slash.
}
