% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.utils.R
\name{bin}
\alias{bin}
\title{Bin a variable in a data frame
## The data frame is expanded by the bin center and bin width.  "var"}
\usage{
bin(df, var, bins)
}
\arguments{
\item{df}{Data frame.  Evaluation occurs with \code{df} as the
environment.}

\item{var}{Expression.  The expression is lazily evaluated.  The
result of the evaluation is then binned.}

\item{bins}{A single number, a vector of numbers, or an expression.
If a single number, \code{pretty(n = bins)} is used on the
result of evaluating \code{var} to determine bins; if a vector,
they are taken as bin edges; if an expression, it is lazily
evaluated and the result used as bin edges.}
}
\value{
A data frame that contains \code{df} and two new variables,
    \code{x_bin} indicating which bin the observation falls into
    and \code{x_width} indicating the width of that bin.
}
\description{
Bin a variable in a data frame
## The data frame is expanded by the bin center and bin width.  "var"
}

