% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_funcs.R
\name{create_dir_if_not}
\alias{create_dir_if_not}
\title{Create a directory if it doesn't exist}
\usage{
create_dir_if_not(d, recursive = T)
}
\arguments{
\item{d}{A path to the location of a directory}

\item{recursive}{Logical. If set to true all the recursive locations will
be created as well.}
}
\description{
Creates a directory after making sure it doesn't already exist.
}
