% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.utils.R
\name{dist.gc}
\alias{dist.gc}
\title{Great-circle distances}
\usage{
dist.gc(lon1, lon2, lat1, lat2)
}
\arguments{
\item{lon1}{Vector of longitudes (in degrees)}

\item{lon2}{Vector of longitudes (in degrees)}

\item{lat1}{Vector of latitudes (in degrees)}

\item{lat2}{Vector of latitudes (in degrees)}
}
\value{
Vector of distances (in km) between points
}
\description{
Calculate great-circle distances based on the haversin formula
}
\examples{

}
