% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plots.R
\name{doPCAMultiPlot}
\alias{doPCAMultiPlot}
\title{Plot a PCA multi-scoreplot or biplot.}
\usage{
doPCAMultiPlot(
  PC,
  nComps,
  samp_info = NULL,
  col = NULL,
  shape = NULL,
  labs = F,
  topNFeats = NULL,
  biplot = F,
  point_size = 3
)
}
\arguments{
\item{PC}{A \code{prcomp} object.}

\item{nComps}{Integer. The number of components to be plotted.}

\item{samp_info}{A dataframe with metadata of the samples, for coloring.
Needs to have a column named \code{"sample"} containing sample names of the
samples in \code{prcomp} object. If not provided, no shaping or coloring of the
points will be possible.}

\item{col}{A variable in \code{samp_info} dataframe which will be used to color
the points.
#' @param shape A variable in \code{samp_info} dataframe which will be used to
shape the points.}

\item{labs}{Logical. If sample labels should be displayed in the plot.}

\item{topNFeats}{Integer. Number of top contributing components to be
displayed in the biplot. By default all features will be displayed.}

\item{point_size}{Integer. Size of the points for the score plot.}
}
\value{
A \code{ggplot} object of the scoreplot/biplot.
}
\description{
Given a \code{prcomp} object, and the number of PCs to be plotted, plots either a
multi-scoreplot or a multi-biplot, with several options.
}
