% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.utils.R
\name{geom_world_polygon}
\alias{geom_world_polygon}
\title{World map outline}
\usage{
geom_world_polygon(col = "black", fill = NA, lwd = 0.5, ...)
}
\arguments{
\item{col}{Line color}

\item{fill}{Fill color}

\item{lwd}{Line width}

\item{...}{Other arguments passed to \code{\link{geom_polygon}}}
}
\value{
A \code{geom_polygon} which can be added to a \code{ggplot}
}
\description{
This functions returns the outline of the continents and major
islands, in lat/lon projection, WGS84 datum, 1:110M resolution,
from Natural Earth, as a \code{\link{geom_polygon}} that can be
added to a \code{\link{ggplot}}.
}

