% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plots.R
\name{getTopContrib}
\alias{getTopContrib}
\title{Obtain the top contributing variables to two PCs}
\usage{
getTopContrib(PC, topN = 12, x = "PC1", y = "PC2")
}
\arguments{
\item{PC}{A \code{prcomp} object.}

\item{topN}{Integer. The number of top contributing variables to be returned.}

\item{x}{The first PC for obtaining top contributing variables.}

\item{y}{The second PC for obtaining top contributing variables.}
}
\value{
A character vector of the top n contributing variables.
}
\description{
Given a \code{prcomp} object and two PCs, gives the n top contributing variables
to those two PCs, based on the variance contributed to the PCs by each
variable.
}
