% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.utils.R
\name{nc.to.df}
\alias{nc.to.df}
\title{Read a list of variables from a NetCDF file and return them as a data.frame}
\usage{
nc.to.df(nc, vars, spread = TRUE, na.rm = FALSE, mask)
}
\arguments{
\item{nc}{ncdf4 NetCDF object}

\item{vars}{Vector of variable names}

\item{spread}{Spread variables into columns?}

\item{na.rm}{Remove NA input rows?}

\item{mask}{(Optional) logical vector; FALSE means discard the corresponding data value}
}
\value{
data.frame of NetCDF contents as vectors with variable
    names as column names
}
\description{
Read a list of variables from a NetCDF file and return them as a data.frame
}
\examples{

}
