% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plots.R
\name{plotPCA}
\alias{plotPCA}
\title{Plot a PCA scoreplot or biplot.}
\usage{
plotPCA(
  PC,
  x = "PC1",
  y = "PC2",
  samp_info = NULL,
  col = NULL,
  shape = NULL,
  labs = NULL,
  biplot = F,
  topNFeats = NULL,
  fix_coord = T,
  point_size = 3
)
}
\arguments{
\item{PC}{A \code{prcomp} object.}

\item{x}{The principal component to be shown in the x axis. \code{"PCN"}.}

\item{y}{The principal component to be shown in the y axis. \code{"PCN"}.}

\item{samp_info}{A dataframe with metadata of the samples, for coloring.
Needs to have a column named \code{"sample"} containing sample names of the
samples in \code{prcomp} object. If not provided, no shaping or coloring of the
points will be possible.}

\item{col}{A variable in \code{samp_info} dataframe which will be used to color
the points.
#' @param shape A variable in \code{samp_info} dataframe which will be used to
shape the points.}

\item{labs}{The samples whose labels should be displayed in the plot,
introduced as a character vector. Defaults to no labels. If set to \code{"all"},
all samples will be labeled.}

\item{topNFeats}{Integer. Number of top contributing components to be
displayed in the biplot. By default all features will be displayed.}

\item{fix_coord}{Logical. If fixed coordinate scale of units in both axes
should be forced.}

\item{point_size}{Integer. Size of the points for the score plot.}
}
\value{
A \code{ggplot} object of the scoreplot/biplot.
}
\description{
Given a \code{prcomp} object, plots either a scoreplot or a biplot, with several
options.
}
