% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotutils-package.R
\docType{package}
\name{plotutils-package}
\alias{plotutils}
\alias{plotutils-package}
\title{plotutils: Setting a custom ggplot2 theme}
\description{
A package created to set my own custom ggplot2 theme globally. It is built upon the ggplot2 theme_minimal() and rearranges title, legend, axes etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/L-Groeninger/plotutils}
  \item Report bugs at \url{https://github.com/L-Groeninger/plotutils/issues}
}

}
\author{
\strong{Maintainer}: Lukas Gröninger \email{lukas.groeninger@gmail.com}

}
\keyword{internal}
