% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{quant_norm}
\alias{quant_norm}
\title{Quantile normalize}
\usage{
quant_norm(m, axis = 2, train_means = NULL)
}
\arguments{
\item{m}{A matrix or dataframe with the data to be normalized}

\item{axis}{If normalization will be performed column-wise (default) or
row-wise.}

\item{train_means}{A vector of rank means.}
}
\description{
Performs quantile normalization in the input data.
}
