% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.utils.R
\name{remap_unstructured}
\alias{remap_unstructured}
\title{Interpolate "unstructured" data frame onto lon/lat grid}
\usage{
remap_unstructured(df.src, df.dest, fill = 0)
}
\arguments{
\item{df.src}{Data frame.  Specifies the data to be interpolated.
Must contain the variables \code{lon} and \code{lat}.}

\item{df.dest}{Data frame.  Specifies the lon/lat grid to be used
for interpolation.  Must contain the variables \code{lon} and
\code{lat}.}

\item{fill}{Numeric.  Fill NA rows with this value, as NA is not
allowed in the interpolation functions.}
}
\value{
A data frame that contains the data from df.src
    interpolated onto the df.dest grid.
}
\description{
Arguments are two data frames.  The source is on an "unstructured"
grid, i.e., one not rectangular in lon/lat space.  The destination
is a lon/lat grid, i.e., of the form \code{df.src <-
mutate(expand.grid(lon = lon, lat = lat), ...)}
}
