% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize.R
\name{sanitize.numbers}
\alias{sanitize.numbers}
\title{Sanitize numbers for tikzDevice}
\usage{
sanitize.numbers(
  str,
  type,
  math.style.negative = FALSE,
  math.style.exponents = FALSE
)
}
\arguments{
\item{str}{Character string to sanitize}

\item{type}{Better choose "latex" for this one, or you will be
disappointed}

\item{math.style.negative}{Boolean that controls whether negative
signs are typeset as negative signs (TRUE is the right answer
here)}

\item{math.style.exponents}{Controls whether exponential notation
is set in LaTeX math mode (TRUE) or ensurmath mode
("ensuremath") or UTF-8 ("UTF8")}
}
\value{
A sanitized string that will look nice and mathy when
    typeset with LaTeX
}
\description{
Adapted from \code{xtable}; regexps better handle conversion to
exponent notation
}
