% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plots.R
\name{save_plot}
\alias{save_plot}
\title{Save the plot}
\usage{
save_plot(filename, plot, width = NA, height = NA)
}
\arguments{
\item{filename}{The path to the file to save the plot}

\item{plot}{A ggplot object}

\item{width}{With of the plot}

\item{height}{Height of the plot}
}
\description{
Wrapper function of \code{ggsave}.
}
