% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.utils.R
\name{scale_geo}
\alias{scale_geo}
\alias{scale_x_geo}
\alias{scale_y_geo}
\alias{scale_x_geo_zonmean}
\title{Geographic scales (with degree labels)}
\usage{
scale_x_geo(facet = FALSE)

scale_y_geo()

scale_x_geo_zonmean()
}
\arguments{
\item{facet}{Boolean.  Will the plots be faceted?  On faceted
plots, the rightmost degree label is omitted to prevent
overplotting.}
}
\value{
A \code{scale_x/y_continuous} that can be added to a \code{ggplot}
}
\description{
Geographic scales (with degree labels)
}
\section{Functions}{
\itemize{
\item \code{scale_x_geo}: longitude (x) axis

\item \code{scale_y_geo}: latitude (y) axis

\item \code{scale_x_geo_zonmean}: latitude (x) axis for zonal-mean plots
}}

