% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_custom_theme.R
\name{set_custom_theme}
\alias{set_custom_theme}
\title{Applying the custom theme function}
\usage{
set_custom_theme(base_size = 14)
}
\arguments{
\item{base_size}{Defining the base size of the font.}
}
\description{
This function is just a wrapper around the theme_barlow() function.
It applies the theme globally and makes sure that the
required "Barlow" font is installed.
}
\examples{
set_custom_theme(base_size = 16)

}
