% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{stand}
\alias{stand}
\title{Standarize data}
\usage{
stand(m, axis = 2, scale = T, center = T)
}
\arguments{
\item{m}{A matrix or dataframe with the data to be normalized}

\item{axis}{Numeric. \code{1} or \code{2}. If normalization will be performed
column-wise (\code{2}, default) or row-wise (\code{1}).}

\item{scale}{Logical. If data should be scaled (divided by sd).}

\item{center}{Logical. If data should be centered (substract the mean).}
}
\description{
Performs z-score standarization. Supports standarization when the input
matrix cntains \code{NA}s, computing the standardized matrix excluding them.
}
