% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_custom_theme.R
\name{theme_barlow}
\alias{theme_barlow}
\title{Barlow theme function}
\usage{
theme_barlow(base_size = 14)
}
\arguments{
\item{base_size}{Defining the base size of the font.}
}
\description{
This function builds upon the ggplot2::theme_minimal() theme
and adds/rearranges some grid elements.
}
\examples{
library(ggplot2)
library(showtext)
showtext_auto() # Necessary for certain graphics devices to display the custom font

ggplot(mtcars, aes(wt, mpg)) + geom_point() + theme_barlow()

}
