% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_fast.R
\name{write_table_fast}
\alias{write_table_fast}
\title{Write table faster}
\usage{
write_table_fast(
  df,
  f,
  row.names = T,
  col.names = T,
  sep = ",",
  sep2 = c("", "|", ""),
  quote = "auto"
)
}
\arguments{
\item{df}{A \code{data.frame}.}

\item{f}{A path to the file to be written.}

\item{row.names}{Logical. If row names should be written in the file.}

\item{col.names}{Logical. If column names should be written in the file.}

\item{sep}{Separator between columns. Defaults to \code{","}.}

\item{sep2}{Separator within columns. Defaults to \code{c("","|","")}. sep2\link{1} is
written at the start of the output field, sep2\link{2} is placed between each item
and sep2\link{3} is written at the end. sep2\link{1} and sep2\link{3} may be any length
strings including empty "" (default). sep2\link{2} must be a single character and
(when list columns are present and therefore sep2 is used) different from
both sep and dec.}

\item{quote}{When "auto", character fields, factor fields and column names
will only be surrounded by double quotes when they need to be; i.e., when the
field contains the separator sep, a line ending \n, the double quote itself
or (when list columns are present) sep2\link{2} (see sep2 below). If FALSE the
fields are not wrapped with quotes even if this would break the CSV due to
the contents of the field. If TRUE double quotes are always included other
than around numeric fields, as write.csv.}
}
\value{
A \code{data.frame} with the contents of the file \code{f}.
}
\description{
Wrapper function to data.table::fwrite, to work as write.csv, but faster
}
