% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.utils.R
\name{geom_world_polygon}
\alias{geom_world_polygon}
\title{World map outline}
\usage{
geom_world_polygon(col = "black", fill = NA, lwd = 0.5, highres = FALSE, ...)
}
\arguments{
\item{col}{Line color}

\item{fill}{Fill color}

\item{lwd}{Line width}

\item{highres}{Logical.  High resolution (1:10m) or default (1:110m)}

\item{...}{Other arguments passed to \code{\link{geom_polygon}}}
}
\value{
A \code{geom_polygon} which can be added to a \code{ggplot}
}
\description{
This functions returns the outline of the continents and major
islands, in lat/lon projection, WGS84 datum, 1:110M resolution,
from Natural Earth, as a \code{\link{geom_polygon}} that can be
added to a \code{\link{ggplot}}.
}
\examples{
expand.grid(lon = -180:180, lat = -90:90) \%>\%
    mutate(fill = cos(pi * lat / 180)) \%>\%
    ggplot(aes(x = lon, y = lat, fill = fill)) +
    geom_raster() +
    geom_world_polygon() +
    scale_x_geo() + scale_y_geo()
}
